server/requirements.txt
# Flask Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-CORS>=4.0.0

# Database
SQLAlchemy>=2.0.0

# Cloud Services
supabase>=2.0.0
boto3>=1.34.0

# Development Tools
pytest>=7.4.0
pytest-flask>=1.3.0
black>=23.0.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
